requirements.txt
dspy
python-dotenv
transitions
asyncio
